ethix-ai-sidebar/pyproject.toml
[project]
name = "ethix-ai-sidebar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "groq>=0.37.1",
    "langchain>=1.2.12",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
]
